// dashboard-ts/src/api-types.d.ts
export type GetDeviceByID = {
	id: number;
	is_assigned: boolean;
	name: string;
	nfc_data: string;
	timestamp: string;
	expires_at: string;
	device_type: string;
	latitude: number;
	longitude: number;
	altitude: number;
	radius_in_meters: number;
	circle_geojson: {
		type: string;
		coordinates: Array<Array<Array<number>>>;
	};
	auto_update_circle_geojson: boolean;
	created_by: number;
	assigned_to_bandobas: number;
};